Ahmad

+++ src/main/java/com/example/hotelmanagementsystem/Entity/Customer.java
@@ -10,14 +10,13 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-//hi hi
 @Table(name = "Customer")
 public class Customer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "name" , nullable = false)
     private String name;
